requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
einops==0.8.2
filelock==3.25.2
fonttools==4.62.1
fsspec==2026.2.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
idna==3.11
iopath==0.1.10
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
librt==0.8.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mcp==1.26.0
mdurl==0.1.2
mpmath==1.3.0
mypy==1.20.0
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.3
ollama==0.6.1
opencv-python==4.13.0.92
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
polars==1.39.3
polars-runtime-32==1.39.3
portalocker==3.2.0
psutil==7.2.2
pycocotools==2.0.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2026.3.32
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
sam3 @ file:///C:/Users/rr006/projects/segmentiq/sam3
scipy==1.17.1
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sse-starlette==3.3.4
starlette==1.0.0
sympy==1.14.0
timm==1.0.26
torch==2.11.0+cu128
torchvision==0.26.0+cu128
tqdm==4.67.3
triton-windows==3.6.0.post26
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.33
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
